Breaking change: blacklist_threshold receives AR-scope.
Previously it only received a count of the bounces, by giving it a scope
it's more flexible in determining the threshold (ie. only check bounces
of the last 2 weeks).

The default threshold is also changes: only block emails where there
have been more then 10 blocks in the past 2 weeks. Perhaps this is
default is *too* loose, I'm not sure yet. The previous default was too
strict for our use-case.